add post to act.js
