+++ api/entries.json
@@ -36,11 +36,11 @@
       "id": 7
     },
     {
-      "date": "2019-06-27",
-      "title": "hey",
-      "entry": "man",
+      "date": "2019-07-14",
+      "title": "react",
+      "entry": "Im'm a react boy",
       "mood": "Okay",
-      "id": 8
+      "id": 9
     }
   ]
 }
