+++ src/scripts/act.js
@@ -1,3 +1,5 @@
+import { API } from "./data.js"
+
 export const act = {
   filterButtons() {
     let allRadioButtons = document.getElementsByName('mood')
@@ -14,5 +16,36 @@ export const act = {
         });
       })
     })
+  },
+
+  postNewEntry() {
+    document.querySelector("#submit").addEventListener("click", () => {
+      function createEntryObj(date, title, entry, mood) {
+        return {
+          date: date,
+          title: title,
+          entry: entry,
+          mood: mood
+        };
+      }
+      let currentObj = createEntryObj(
+        document.querySelector("#journalDate").value,
+        document.querySelector("#concepts").value,
+        document.querySelector("#entry").value,
+        document.querySelector("#Mood").value
+      );
+
+      function formCheck(object) {
+        return Object.values(object).includes("")
+      }
+
+      if (formCheck(currentObj)) {
+        alert("All fields must be filled out!")
+      }
+      else {
+        API.postEntry(currentObj)
+      }
+
+    });
   }
 }
