+++ src/scripts/components/searchbar.js
@@ -1,0 +1,3 @@
+export const searchBar = {
+
+}
